# docs/api-documentation.yaml
openapi: 3.0.3
info:
  title: SubscriptionPro API
  description: |
    Comprehensive API for the SubscriptionPro subscription management platform.
    
    This API provides endpoints for managing users, subscriptions, products, payments, and administrative functions.
    
    ## Authentication
    
    The API uses JWT (JSON Web Tokens) for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    
    API requests are rate limited to prevent abuse:
    - General endpoints: 200 requests per day, 50 per hour
    - Authentication endpoints: 5 requests per minute
    
    ## Error Handling
    
    The API returns standard HTTP status codes and JSON error responses:
    ```json
    {
      "error": "Error message",
      "code": "ERROR_CODE",
      "details": {}
    }
    ```
    
  version: 1.0.0
  contact:
    name: SubscriptionPro Support
    email: support@subscriptionpro.com
    url: https://subscriptionpro.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.subscriptionpro.com
    description: Production server
  - url: https://staging-api.subscriptionpro.com
    description: Staging server
  - url: http://localhost:5000
    description: Development server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
          description: Unique user identifier
        email:
          type: string
          format: email
          description: User's email address
        first_name:
          type: string
          description: User's first name
        last_name:
          type: string
          description: User's last name
        phone_number:
          type: string
          description: User's phone number
        date_of_birth:
          type: string
          format: date
          description: User's date of birth
        user_role:
          type: string
          enum: [user, admin]
          description: User's role in the system
        address_line1:
          type: string
          description: Primary address line
        address_line2:
          type: string
          description: Secondary address line
        city:
          type: string
          description: City
        state:
          type: string
          description: State or province
        postal_code:
          type: string
          description: Postal or ZIP code
        country:
          type: string
          description: Country
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
      required:
        - email
        - first_name
        - last_name
...
    Subscription:
      type: object
      properties:
        subscription_id:
          type: string
          format: uuid
          description: Unique subscription identifier
        user_id:
          type: string
          format: uuid
          description: User who owns the subscription
        product_id:
          type: string
          format: uuid
          description: Subscribed product
        status:
          type: string
          enum: [active, paused, canceled]
          description: Subscription status
        frequency:
          type: string
          enum: [weekly, monthly, quarterly, yearly]
          description: Delivery frequency
        quantity:
          type: integer
          minimum: 1
          description: Quantity per delivery
        amount:
          type: number
          format: decimal
          description: Subscription amount
        currency:
          type: string
          default: INR
        start_date:
          type: string
          format: date
          description: Subscription start date
        end_date:
          type: string
          format: date
          description: Subscription end date (if applicable)
        next_delivery_date:
          type: string
          format: date
          description: Next scheduled delivery
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - user_id
        - product_id
        - frequency
        - amount
        - start_date
...
    AuditLog:
      type: object
      properties:
        log_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        action_type:
          type: string
        entity_type:
          type: string
        entity_id:
          type: string
          format: uuid
        details:
          type: string
        ip_address:
          type: string
        user_agent:
          type: string
        timestamp:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
...
paths:
...
  /subscriptions:
    get:
      tags:
        - Subscriptions
      summary: List user subscriptions
      description: Retrieve all subscriptions for the current user
      parameters:
        - name: status
          in: query
          description: Filter by subscription status
          schema:
            type: string
            enum: [active, paused, canceled]
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Subscriptions retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscriptions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Subscription'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer

    post:
      tags:
        - Subscriptions
      summary: Create subscription
      description: Create a new subscription for a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                product_id:
                  type: string
                  format: uuid
                frequency:
                  type: string
                  enum: [weekly, monthly, quarterly, yearly]
                quantity:
                  type: integer
                  minimum: 1
                start_date:
                  type: string
                  format: date
              required:
                - product_id
                - frequency
                - start_date
      responses:
        '201':
          description: Subscription created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /subscriptions/{subscription_id}/pause:
    post:
      tags:
        - Subscriptions
      summary: Pause subscription
      description: Pause an active subscription
      parameters:
        - name: subscription_id
          in: path
          required: true
          description: Subscription ID
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Reason for pausing
                resume_date:
                  type: string
                  format: date
                  description: Optional resume date
      responses:
        '200':
          description: Subscription paused successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '400':
          description: Cannot pause subscription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /subscriptions/{subscription_id}/resume:
    post:
      tags:
        - Subscriptions
      summary: Resume subscription
      description: Resume a paused subscription
      parameters:
        - name: subscription_id
          in: path
          required: true
          description: Subscription ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Subscription resumed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '400':
          description: Cannot resume subscription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /subscriptions/{subscription_id}/cancel:
    post:
      tags:
        - Subscriptions
      summary: Cancel subscription
      description: Cancel an active or paused subscription
      parameters:
        - name: subscription_id
          in: path
          required: true
          description: Subscription ID
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Reason for cancellation
                immediate:
                  type: boolean
                  default: false
                  description: Cancel immediately or at end of current period
      responses:
        '200':
          description: Subscription canceled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '400':
          description: Cannot cancel subscription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /admin/users:
    get:
      tags:
        - Admin
      summary: List all users (Admin)
      description: Retrieve a list of all users (admin only)
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          description: Search users by email or name
          schema:
            type: string
        - name: role
          in: query
          description: Filter by user role
          schema:
            type: string
            enum: [user, admin]
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/subscriptions:
    get:
      tags:
        - Admin
      summary: List all subscriptions (Admin)
      description: Retrieve a list of all subscriptions (admin only)
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          description: Filter by subscription status
          schema:
            type: string
            enum: [active, paused, canceled]
        - name: user_email
          in: query
          description: Filter by user email
          schema:
            type: string
      responses:
        '200':
          description: Subscriptions retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscriptions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Subscription'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
...
  /admin/audit-logs:
    get:
      tags:
        - Admin
      summary: Get audit logs (Admin)
      description: Retrieve system audit logs (admin only)
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 50
        - name: action_type
          in: query
          description: Filter by action type
          schema:
            type: string
        - name: entity_type
          in: query
          description: Filter by entity type
          schema:
            type: string
        - name: user_id
          in: query
          description: Filter by user ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Audit logs retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLog'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer

  # Health Check Endpoint
...